@@ -7,21 +7,19 @@
     "test": "jest"
   },
   "dependencies": {
+    "@expo/vector-icons": "^5.0.0",
     "axios": "^0.15.3",
-    "firebase": "^3.6.9",
+    "expo": "^17.0.0",
     "material-design-icons": "^3.0.1",
-    "native-base": "2.1.0",
-    "react": "~15.4.0-rc.4",
-    "react-native": "^0.41.2",
+    "native-base": "2.1.3",
+    "react": "16.0.0-alpha.6",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-17.0.0.tar.gz",
     "react-native-elements": "^0.10.2",
-    "react-native-keyboard-aware-scrollview": "^1.1.5",
-    "react-native-lock": "^0.5.0",
+    "react-native-google-analytics-bridge": "^5.0.1",
     "react-native-menu": "^0.20.2",
-    "react-native-vector-icons": "^4.0.1",
-    "react-navigation": "^1.0.0-beta.7",
+    "react-navigation": "^1.0.0-beta.11",
     "react-redux": "^5.0.1",
     "redux": "^3.6.0",
-    "redux-actions": "^1.2.0",
     "redux-thunk": "^2.1.0"
   },
   "devDependencies": {
@@ -38,5 +36,6 @@
   },
   "jest": {
     "preset": "react-native"
-  }
+  },
+  "main": "main.js"
 }
