@@ -45,6 +45,6 @@
     "eslint-plugin-jsx-a11y": "^4.0.0",
     "eslint-plugin-react": "^6.10.3",
     "nodemon": "^1.11.0",
-    "webpack": "^2.3.1"
+    "webpack": "^2.6.1"
   }
 }
