@@ -1,5 +1,6 @@
-export * from './FeedbackActions';
-export * from './AuthActions';
-export * from './SolutionActions';
-export * from './actions_user';
+export * from './actions_auth';
 export * from './actions_features';
+export * from './actions_feedback';
+export * from './actions_projects';
+export * from './actions_solutions';
+export * from './actions_user';
