@@ -4,9 +4,12 @@ export const ADD_PROJECT_UPVOTE = 'add_project_upvote';
 export const REMOVE_PROJECT_UPVOTE = 'remove_project_upvote';
 export const LOAD_PROJECT_UPVOTES = 'load_project_upvotes';
 export const LOAD_SOLUTION_UPVOTES = 'load_solution_upvotes';
+export const ADD_PROJECT_DOWNVOTE = 'add_project_downvote';
+export const REMOVE_PROJECT_DOWNVOTE = 'remove_project_downvote';
+export const LOAD_PROJECT_DOWNVOTES = 'load_project_downvotes';
+export const LOAD_SOLUTION_DOWNVOTES = 'load_solution_downvotes';
 export const SAVE_PROJECT_CHANGES = 'save_project_changes';
 export const ADD_PROJECT = 'add_project';
-export const DELETE_PROJECT = 'delete_project';
 export const REQUESTED_PROJECTS = 'requested_projects';
 export const RECEIVED_PROJECTS = 'received_projects';
 export const SUBMIT_FEEDBACK = 'submit_feedback';
@@ -22,7 +25,6 @@ export const SENT_AUTHORIZATION_EMAIL_FAIL = 'sent_authorization_email_fail';
 export const AUTHORIZING_USER = 'authorizing_user';
 export const AUTHORIZE_USER_SUCCESS = 'authorize_user_success';
 export const AUTHORIZE_USER_FAIL = 'authorize_user_fail';
-export const LOAD_STATE_SUCCESS = 'load_state_success';
 export const EMAIL_SAVE_SUCCESS = 'email_save_success';
 export const TOKEN_SAVE_SUCCESS = 'token_save_success';
 export const LOG_OUT_USER = 'log_out_user';
@@ -35,6 +37,8 @@ export const SUBMIT_SOLUTION_SUCCESS = 'submit_solution_success';
 export const SUBMIT_SOLUTION_FAIL = 'submit_solution_fail';
 export const ADD_SOLUTION_UPVOTE = 'add_solution_upvote';
 export const REMOVE_SOLUTION_UPVOTE = 'remove_solution_upvote';
+export const ADD_SOLUTION_DOWNVOTE = 'add_solution_downvote';
+export const REMOVE_SOLUTION_DOWNVOTE = 'remove_solution_downvote';
 export const SAVE_SOLUTION_CHANGES = 'save_solution_changes';
 export const ADD_SOLUTION_TO_STATE = 'add_solution_to_state';
 
