@@ -5,11 +5,10 @@ import { StackNavigator } from 'react-navigation';
 import { Icon } from 'react-native-elements';
 
 // Import Scenes
-import Feedback from '../scenes/Feedback';
+import FeedbackSubmit from '../scenes/FeedbackSubmit';
 import Settings from '../scenes/Settings';
 import Submitted from '../scenes/Submitted';
 import styles from '../styles/common/navStyles';
-
 
 function settingsButton(navigate) {
   const right = (
@@ -24,31 +23,53 @@ function settingsButton(navigate) {
   return right;
 }
 
-const FeedbackStack = StackNavigator(
+// Stack of scenes
+const scenes = StackNavigator(
   {
-    Feedback: {
-      screen: Feedback,
+    FeedbackSubmit: {
+      screen: FeedbackSubmit,
       navigationOptions: ({ navigation }) => ({
         title: 'Submit Feedback',
         headerRight: settingsButton(navigation.navigate),
-        headerStyle: { height: styles.header.height },
+        headerStyle: {
+          height: styles.header.height,
+          marginTop: styles.header.marginTop,
+        },
       }),
     },
     Settings: {
       screen: Settings,
       navigationOptions: {
         title: 'Settings',
-        headerStyle: { height: styles.header.height },
+        headerStyle: {
+          height: styles.header.height,
+          marginTop: styles.header.marginTop,
+        },
       },
     },
     Submitted: {
       screen: Submitted,
       navigationOptions: {
         title: 'Feedback Received',
-        headerStyle: { height: styles.header.height },
+        headerStyle: {
+          height: styles.header.height,
+          marginTop: styles.header.marginTop,
+        },
       },
     },
   },
 );
 
-export default FeedbackStack;
+// Stack options
+const options = {
+  tabBarLabel: 'Submit Feedback',
+  tabBarIcon: ({ tintColor }) => <Icon name="create" size={22} color={tintColor} />,
+  cardStack: { gesturesEnabled: false },
+};
+
+const FeedbackSubmitStack = {
+  screen: scenes,
+  navigationOptions: options,
+};
+
+export default FeedbackSubmitStack;
