@@ -11,7 +11,7 @@ async function loadToken() {
     // AsyncStorage.removeItem(`${ROOT_STORAGE}token`);
     const token = await AsyncStorage.getItem(`${ROOT_STORAGE}token`) || null;
     if (token === null) {
-      store.dispatch(actions.authorizeFail());
+      store.dispatch(actions.authorizeUserFail());
     } else {
       store.dispatch(actions.pullProjects(token));
       store.dispatch(actions.pullSolutions(token));
@@ -37,10 +37,25 @@ async function loadUpvotes() {
     console.log(error);
   }
 }
+async function loadDownvotes() {
+  try {
+    // Project Upvotes
+    let projectDownvotes = await AsyncStorage.getItem(`${ROOT_STORAGE}downvotes`) || '[]';
+    projectDownvotes = JSON.parse(projectDownvotes);
+    store.dispatch(actions.loadProjectDownvotes(projectDownvotes));
+
+    // Solution Downvotes
+    let solutionDownvotes = await AsyncStorage.getItem(`${ROOT_STORAGE}solutionDownvotes`) || '[]';
+    solutionDownvotes = JSON.parse(solutionDownvotes);
+    store.dispatch(actions.loadSolutionDownvotes(solutionDownvotes));
+  } catch (error) {
+    console.log(error);
+  }
+}
 
 async function loadDoNotDisplayList() {
   try {
-    //AsyncStorage.removeItem(`${ROOT_STORAGE}doNotDisplayList`);
+    // AsyncStorage.removeItem(`${ROOT_STORAGE}doNotDisplayList`);
     let doNotDisplayList = await AsyncStorage.getItem(`${ROOT_STORAGE}doNotDisplayList`) || '[]';
     doNotDisplayList = JSON.parse(doNotDisplayList);
     store.dispatch(actions.loadDoNotDisplayList(doNotDisplayList));
@@ -60,17 +75,28 @@ async function loadInstructions() {
   }
 }
 
+async function clearAsyncStorage() {
+  try {
+    await AsyncStorage.removeItem(`${ROOT_STORAGE}token`);
+    await AsyncStorage.removeItem(`${ROOT_STORAGE}upvotes`);
+    await AsyncStorage.removeItem(`${ROOT_STORAGE}solutionUpvotes`);
+    await AsyncStorage.removeItem(`${ROOT_STORAGE}downvotes`);
+    await AsyncStorage.removeItem(`${ROOT_STORAGE}solutionDownvotes`);
+    await AsyncStorage.removeItem(`${ROOT_STORAGE}doNotDisplayList`);
+    await AsyncStorage.removeItem(`${ROOT_STORAGE}instructionsViewed`);
+  } catch (error) {
+    console.log(error);
+  }
+}
+
 // Initialize saved state
 const loadOnLaunch = () => {
-  Promise.all([
-    loadUpvotes(),
-    loadDoNotDisplayList(),
-    loadToken(),
-    loadInstructions(),
-  ])
-  .then(() => {
-    store.dispatch(actions.loadStateSuccess());
-  });
+  // clearAsyncStorage();
+  loadUpvotes();
+  loadDoNotDisplayList();
+  loadToken();
+  loadInstructions();
+  loadDownvotes();
 };
 
 export default loadOnLaunch;
