@@ -7,12 +7,10 @@ import {
   AUTHORIZE_USER_SUCCESS,
   AUTHORIZE_USER_FAIL,
   LOAD_TOKEN,
-  LOAD_STATE_SUCCESS,
   LOG_OUT_USER,
 } from '../actions/types';
 
 const INITIAL_STATE = {
-  loadingState: true,
   email: '',
   token: null,
   loading: false,
@@ -23,8 +21,6 @@ const INITIAL_STATE = {
 
 export default (state = INITIAL_STATE, action) => {
   switch (action.type) {
-    case LOAD_STATE_SUCCESS:
-      return { ...state, loadingState: false };
     case SENDING_AUTHORIZATION_EMAIL:
       return { ...state, loading: true };
     case SENT_AUTHORIZATION_EMAIL_SUCCESS:
@@ -40,7 +36,7 @@ export default (state = INITIAL_STATE, action) => {
     case LOAD_TOKEN:
       return { ...state, token: action.payload };
     case LOG_OUT_USER:
-      return { ...state, loggedIn: false, token: null };
+      return INITIAL_STATE;
     default:
       return state;
   }
