@@ -2,7 +2,7 @@
 import { combineReducers } from 'redux';
 
 // Import reducers
-import main from './reducer_main';
+import feedback from './reducer_feedback';
 import projects from './reducer_projects';
 import auth from './reducer_auth';
 import user from './reducer_user';
@@ -10,7 +10,7 @@ import solutions from './reducer_solutions';
 import features from './reducer_features';
 
 export default combineReducers({
-  main,
+  feedback,
   projects,
   solutions,
   auth,
