@@ -2,7 +2,6 @@
 import React, { Component } from 'react';
 import { Text, View, Keyboard, TouchableWithoutFeedback, Image, TouchableOpacity, StyleSheet } from 'react-native';
 import { connect } from 'react-redux';
-import { NavigationActions } from 'react-navigation';
 
 // Import components and action creators
 import { Card, CardSection, Input, Button, Spinner } from '../components/common';
@@ -10,10 +9,13 @@ import { sendAuthorizationEmail, authorizeUserFail, closeInstructions } from '..
 import styles from '../styles/styles_main';
 import fullScreen from '../../images/backgrounds/EmailInfo.png';
 
+// Import tracking
+// import { tracker } from '../constants';
+
 var styles2 = StyleSheet.create({
   imageContainer: {
     flex: 1,
-    alignItems: 'stretch'
+    alignItems: 'stretch',
   },
   image: {
     flex: 1,
@@ -28,6 +30,8 @@ class SendAuthorizationEmail extends Component {
       email: '',
     };
 
+    // tracker.trackScreenView('Send Auth Email');
+
     this.sendAuthorizationEmail = this.sendAuthorizationEmail.bind(this);
     this.closeInstructions = this.closeInstructions.bind(this);
   }
@@ -35,6 +39,7 @@ class SendAuthorizationEmail extends Component {
   sendAuthorizationEmail() {
     const re = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
     if (re.test(this.state.email)) {
+      Keyboard.dismiss();
       this.props.sendAuthorizationEmail(this.state.email, () => this.props.navigation.navigate('AuthCode'));
     } else {
       this.props.authorizeUserFail('Invalid Email Address');
