@@ -1,6 +1,6 @@
 // Import Libraries
 import React, { Component } from 'react';
-import { View, Keyboard } from 'react-native';
+import { View } from 'react-native';
 import { connect } from 'react-redux';
 import { NavigationActions } from 'react-navigation';
 
@@ -11,7 +11,16 @@ import { logOut } from '../actions';
 import { Button } from '../components/common';
 import styles from '../styles/settings_styles';
 
+// Import tracking
+// import { tracker } from '../constants';
+
 class Settings extends Component {
+  constructor(props) {
+    super(props);
+
+    // tracker.trackScreenViewWithCustomDimensionValues('Settings', { domain: props.features.domain });
+  }
+
   render() {
     const { container } = styles;
 
@@ -23,6 +32,7 @@ class Settings extends Component {
         <View style={{ flex: 1, flexDirection: 'column', alignItems: 'center' }}>
           <Button
             onPress={() => {
+              // tracker.trackEvent('Auth', 'Logged Out', { label: this.props.features.domain });
               this.props.logOut();
               this.props.navigation.navigate('Auth');
             }}
@@ -37,13 +47,14 @@ class Settings extends Component {
 }
 
 Settings.propTypes = {
-  auth: React.PropTypes.object,
   logOut: React.PropTypes.func,
   navigation: React.PropTypes.object,
+  features: React.PropTypes.object,
 };
 
 function mapStateToProps(state) {
-  return state;
+  const { features } = state;
+  return { features };
 }
 
 export default connect(mapStateToProps, { logOut })(Settings);
