@@ -1,25 +1,21 @@
 // Import Libraries
 import React, { Component } from 'react';
-import { View, BackAndroid, Keyboard } from 'react-native';
+import { View, Keyboard } from 'react-native';
 import { connect } from 'react-redux';
 import { NavigationActions } from 'react-navigation';
-
-// Import actions
-import { navigate } from '../actions';
 
 // Import components, functions, and styles
 import { Button } from '../components/common';
 import styles from '../styles/styles_main';
 
-class Submitted extends Component {
-  constructor(props, context) {
-    super(props, context);
+// Import tracking
+// import { tracker } from '../constants';
 
-    BackAndroid.addEventListener('hardwareBackPress', () => {
-      // To do: Update navigation to use react-navigation
-      props.navigate({ type: 'pop' });
-      return true;
-    });
+class Submitted extends Component {
+  constructor(props) {
+    super(props);
+
+    // tracker.trackScreenViewWithCustomDimensionValues('Submitted', { domain: props.features.domain });
   }
 
   componentWillMount() {
@@ -31,7 +27,7 @@ class Submitted extends Component {
       <View style={styles.container}>
         <View style={{ flex: 1, paddingTop: 20 }}>
           {/* To do: To do: Update navigation to use react-navigation */}
-          <Button onPress={() => this.props.navigation.navigate('NewProjects')}>
+          <Button onPress={() => this.props.navigation.navigate('FeedbackSwipe')}>
             Vote on Feedback!
           </Button>
         </View>
@@ -42,10 +38,12 @@ class Submitted extends Component {
 
 Submitted.propTypes = {
   navigation: React.PropTypes.object,
+  features: React.PropTypes.object,
 };
 
 function mapStateToProps(state) {
-  return state;
+  const { features } = state;
+  return { features };
 }
 
-export default connect(mapStateToProps, { navigate })(Submitted);
+export default connect(mapStateToProps)(Submitted);
