@@ -5,7 +5,10 @@ const styles = {
   header: {
     ...Platform.select({
       ios: { height: 60 },
-      android: { height: 45 },
+      android: {
+        height: 45,
+        marginTop: 24,
+      },
     }),
   },
 };
