@@ -1,10 +1,13 @@
 require('dotenv').config();
+
 const express = require('express');
 const mysql = require('mysql');
 const multer = require('multer');
 const jwt = require('jsonwebtoken'); // For authentication
 const bodyParser = require('body-parser'); // For uploading longer/complicated texts
 const aws = require('aws-sdk'); // load aws sdk
+
+aws.config.loadFromPath('config.json'); // load aws config
 const WebClient = require('@slack/client').WebClient; // for Slack
 
 const request = require('request'); // Slack
@@ -12,22 +15,10 @@ const request = require('request'); // Slack
 const app = express();
 const upload = multer(); // for parsing multipart/form-data
 const ses = new aws.SES({ apiVersion: '2010-12-01' }); // load AWS SES
-aws.config.loadFromPath('config.json'); // load aws config
 
 app.use(bodyParser.json()); // for parsing application/json
 app.use(bodyParser.urlencoded({ extended: true })); // for parsing application/x-www-form-urlencoded
 app.use(express.static('public'));
-
-//Slack
-const botToken = process.env.SLACK_API_BOT_TOKEN || '';
-const botWeb = new WebClient(botToken);
-
-// depreciated - admin use only
-const IncomingWebhook = require('@slack/client').IncomingWebhook;
-const token = process.env.SLACK_API_TOKEN || '';
-const web = new WebClient(token);
-const url = process.env.SLACK_WEBHOOK_URL || ''; // see section above on sensitive data
-const webhook = new IncomingWebhook(url);
 
 const connection = mysql.createConnection({
   user: 'root',
@@ -46,6 +37,7 @@ const defaultFromEmail = 'moderator@collaborativefeedback.com';
 
 connection.connect();
 
+// **Mobile App and Website**
 function sendEmail(toEmail, fromEmail, subjectLine, bodyText) {
   ses.sendEmail({
     Source: fromEmail,
@@ -84,238 +76,64 @@ function getDomain(email) {
   return re.exec(email)[0].slice(1);
 }
 
-// Slack
-// Step #1 -- When user clicks 'add to slack', slack passes the API keys to this API
-
-// Saves the timestamp in the database (so we know the key for updating)
-function insertSlot(message, teamId) {
-  connection.query('INSERT INTO slack_slots (team_id, ts) VALUES (?, ?)', [teamId, message.ts], (err) => {
-    if (err) throw err;
-  });
-}
-
-// Posts messages to channel during setup (start with 5)
-function postMessageToSuggestions(text, teamId) {
-  botWeb.chat.postMessage('#suggestions', text, (err, res) => {
-    if (err) {
-      console.log('Error:', err);
-    } else {
-      console.log('message response to suggestions', res);
-      insertSlot(res.message, teamId);
-    }
-  });
-}
-// TO DO -- store API keys in database
-// TO DO -- route the user to the success screen
-// Posts the first 5 messages (requires the admin to have already created the suggestions channel)
-app.get('/slack/auth', (req, res) => {
-  const options = {
-    uri: 'https://slack.com/api/oauth.access?code='
-        + req.query.code +
-        '&client_id=' + process.env.CLIENT_ID +
-        '&client_secret=' + process.env.CLIENT_SECRET,
-    method: 'GET',
-  };
-  request(options, (error, response, body) => {
-    const JSONresponse = JSON.parse(body);
-    if (!JSONresponse.ok) {
-      console.log(JSONresponse);
-      res.send('Error encountered: \n' + JSON.stringify(JSONresponse)).status(200).end();
-    } else {
-      console.log(JSONresponse);
-      res.send('Success!');
-    }
-  });
-});
-
-// Step #2 -- When the user submits a /suggestion
-
-// Gets a list of all the users to send a DM to (for upvoting)
-// TO DO -- switch to users in suggestions channel instead of all users in team (?)
-function getUsers(text, suggestionId, teamId) {
-  botWeb.im.list((err, res) => {
-    if (err) {
-      console.log('ERROR: getUsers ', err);
-    } else {
-      postMessageToDM(text, res.ims, suggestionId, teamId);
-    }
-  });
-}
-
-// depreciated - using callback_id
-function insertDM(dm, suggestionId, teamId) {
-  connection.query('INSERT INTO dms (suggestion_id, ts, team_id, dm_id) VALUES (?, ?, ?, ?)', [suggestionId, dm.ts, teamId, dm.channel], (err) => {
-    if (err) throw err;
-  });
-}
-
-// Sends each user the text of the suggestion and an upvote button
-// TO DO -- automatically turn off notifications for this message
-function postMessageToDM(text, users, suggestionId, teamId) {
-  users.forEach((currentValue) => {
-    botWeb.chat.postMessage(currentValue.id, text, {
-      attachments: [
-        {
-          text: 'Do you agree with this suggestion?',
-          fallback: 'Woops... sorry something went wrong...',
-          callback_id: String(suggestionId),
-          color: '#3AA3E3',
-          attachment_type: 'default',
-          actions: [
-            {
-              name: 'vote',
-              text: 'I agree (vote)',
-              type: 'button',
-              value: 'upvote',
-            },
-            {
-              name: 'skip',
-              text: 'Meh... (skip)',
-              type: 'button',
-              value: 'skip',
-            },
-          ],
-        },
-      ],
-    }, (err, res) => {
-      if (err) {
-        console.log('Error:', err);
-      } else {
-        //console.log('DM Response', res);
-        //insertDM(res, suggestionId, teamId);
-      }
-    });
-  });
-}
-
-// Helper functino for updateBoard -- pull the slack spots for the current team
-function getSlots(teamId) {
-  console.log('get slots');
-  const connectionString = `
-    SELECT
-      ts
-    FROM
-      slack_slots
-    WHERE
-      team_id=?`;
-  connection.query(connectionString, [teamId], (err, slots) => {
-    if (err) throw err;
-    else {
-      getProjects(slots, teamId);
-    }
-  });
-}
-
-// Helper function for updateBoard -- pull the current projects from the database
-function getProjects(slots, teamId) {
-  const connectionString = `
-    SELECT
-      id, text, votes
-    FROM
-      slack_feedback
-    WHERE
-      team_id=?`;
-  connection.query(connectionString, [teamId], (err, feedback) => {
-    if (err) throw err;
-    else {
-      updateBoard(slots, feedback, teamId);
-    }
-  });
-}
-
-// Updates the board to reflect the current status of the database
-// ******* TO DO -- need to save the channel to database too
-function updateBoard(slots, feedback, teamId) {
-  //slotsSorted = slots.sort((a, b) => a.ts - b.ts);
-  feedback.sort((a, b) => a.votes - b.votes).forEach((currentValue, index) => {
-    const text = '*' + currentValue.votes + ' Votes* ' + currentValue.text;
-    const channel = 'C5CSA6ECC';
-    if (index >= slots.length) {
-      postMessageToSuggestions(text, teamId);
-    } else {
-      botWeb.chat.update(slots[index].ts, channel, text, { as_user: true }, (err) => {
-        if (err) {
-          console.log('Error:', err);
-        }
-      });
-    }
-  });
-}
-
-app.post('/slack/suggestion', upload.array(), (req, res) => {
-  // Adds the feedback to the feedback table and updates the board
-  connection.query('INSERT INTO slack_feedback (text, submitted_by, team_id) VALUES (?, ?, ?)', [req.body.text, req.body.user_id, req.body.team_id], (err, res) => {
-    if (err) throw err;
-    else {
-      getSlots(req.body.team_id);
-      // Sends a DM to all users (for upvoting)
-      getUsers(req.body.text, res.insertId, req.body.team_id);
-    }
-  });
-
-  res.sendStatus(200);
-});
-
-// STEP #3 -- When a user 'upvotes'
-app.post('/slack/vote', upload.array(), (req, res) => {
-  const payload = JSON.parse(req.body.payload);
-  const upvote = payload.actions[0].value;
-
-  if (upvote === 'upvote') {
-    // *** TO DO -- updates the votes for that feedback accordingly in the database
-    connection.query('UPDATE slack_feedback SET votes=votes+1 WHERE id=?', [payload.callback_id], (err) => {
-      if (err) throw err;
-      else {
-        getSlots(payload.team.id);
-      }
-    });
-    res.json({
-      response_type: 'ephemeral',
-      replace_original: true,
-      text: '# of Votes increased by 1!',
-    });
-  } else {
-    res.json({
-      response_type: 'ephemeral',
-      replace_original: true,
-      text: 'Suggestion skipped',
-    });
-  }
-});
-
 // Authentication
 app.post('/sendAuthorizationEmail', upload.array(), (req, res) => {
-  const re = /^[a-zA-Z0-9._%+-]+@(?:[a-zA-Z0-9-]+\.)*(?:hbs\.edu|stanford\.edu|gymboree\.com)$/;
+  const re = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
   if (!re.test(req.body.email)) {
-    res.status(400).send('Sorry, your company is not currently supported :(');
+    res.status(400).send('Sorry, this does not appear to be a valid email address :(');
   } else {
     // Step #1: Generate a code
     const code = generatePassword(4);
     console.log(code);
 
-    // Step #2: Add the email, code, and timestamp to the database
-    connection.query('INSERT INTO users (email, passcode) VALUES (?, ?) ON DUPLICATE KEY UPDATE passcode=?, passcode_time=NOW()', [req.body.email, String(code), String(code)], function(err) {
+    // Step #2: Check to see if the user is already in the database
+    let groupId;
+    let connectionString = 'SELECT groupId FROM users WHERE email=?';
+    connection.query(connectionString, [req.body.email], (err, rows) => {
       if (err) throw err;
+      if (!rows.length) {
+        // Step #2A: If not in database see if it has a default groupId
+        connectionString = 'SELECT id FROM features WHERE school=?';
+        connection.query(connectionString, [getDomain(req.body.email)], (err, rows) => {
+          if (err) res.status(400).send('Sorry, this email does not appear to be set up in our system :(');
+          if (!rows.length) {
+            res.status(400).send('Sorry, this email does not appear to be set up in our system :(');
+          } else {
+            sendAuthEmailHelper(req, res, code, rows[0].id);
+          }
+        });
+      } else {
+        sendAuthEmailHelper(req, res, code, rows[0].groupId);
+      }
     });
-
-    if (req.body.email.includes('admin_test')) {
-      res.sendStatus(200);
-    } else {
-      // Step #3: Send an email with the code to the user (make sure it shows up in notification)
-      sendEmail([req.body.email], defaultFromEmail, 'Collaborative Feedback: Verify Your Email Address', 'Enter this passcode: ' + String(code));
-      res.sendStatus(200);
-    }
   }
 });
 
+function sendAuthEmailHelper(req, res, code, groupId) {
+  // Step #3: Add the email, groupId, code, and timestamp to the database
+  connection.query('INSERT INTO users (email, groupId, passcode) VALUES (?, ?, ?) ON DUPLICATE KEY UPDATE passcode=?, groupId=?, passcode_time=NOW()', [req.body.email, groupId, String(code), String(code), groupId], function(err) {
+    if (err) throw err;
+  });
+
+  if (req.body.email.includes('admin_test')) {
+    res.sendStatus(200);
+  } else {
+    console.log('email sent');
+    // Step #3: Send an email with the code to the user (make sure it shows up in notification)
+    sendEmail([req.body.email], defaultFromEmail, 'Collaborative Feedback: Verify Your Email Address', 'Enter this passcode: ' + String(code));
+    res.sendStatus(200);
+  }   
+}
+
 app.post('/authorizeUser', upload.array(), (req, res) => {
   // Step #1: Query the database for the passcode and passcode_time associated with the email address in req.body
-  connection.query('SELECT passcode_time FROM users WHERE email=? AND passcode=?', [req.body.email, req.body.code], (err, rows) => {
-    if (err) throw err;
+  const connectionString = (req.body.code === 'apple') ? 'SELECT groupId FROM users WHERE email=?' : 'SELECT groupId FROM users WHERE email=? AND passcode=?';
+  connection.query(connectionString, [req.body.email, req.body.code], (err, rows) => {
+    if (err) res.status(400).send('Incorrect Code');
     // Step #2: Check that it matches the passcode submitted by the user, if not send error
     // Step #3: If it checks out then create a JWT token and send to the user
-    if (rows.length || req.body.code === 'apple') {
-      const myToken = jwt.sign({ email: req.body.email }, 'buechelejedi16');
+    if (rows.length) {
+      const myToken = jwt.sign({ email: req.body.email, groupId: rows[0].groupId }, 'buechelejedi16');
       res.status(200).json(myToken);
     } else {
       res.status(400).send('Incorrect Code');
@@ -325,12 +143,12 @@ app.post('/authorizeUser', upload.array(), (req, res) => {
 
 app.post('/authorizeAdminUser', upload.array(), (req, res) => {
   // Step #1: Query the database for the passcode and passcode_time associated with the email address in req.body
-  connection.query('SELECT passcode_time FROM users WHERE email=? AND passcode=?', [req.body.email, req.body.code], (err, rows) => {
+  connection.query('SELECT groupId FROM users WHERE email=? AND passcode=?', [req.body.email, req.body.code], (err, rows) => {
     if (err) throw err;
     // Step #2: Check that it matches the passcode submitted by the user, if not send error
     // Step #3: If it checks out then create a JWT token and send to the user
     if ((rows.length || req.body.code === 'apple') && (req.body.adminCode === 'GSB2017' || req.body.adminCode === 'HBS2017' || req.body.adminCode === 'GYM2017')) {
-      const myToken = jwt.sign({ email: req.body.email }, 'buechelejedi16');
+      const myToken = jwt.sign({ email: req.body.email, groupId: rows[0].groupId }, 'buechelejedi16');
       res.status(200).json(myToken);
     } else {
       res.status(400).send('Incorrect Code');
@@ -347,10 +165,10 @@ app.post('/addFeedback', upload.array(), (req, res) => {
       const school = getDomain(decoded.email);
 
       // Send Email
-      const toEmails = ['tyler.hannasch@gmail.com', 'newton1988@gmail.com', 'alicezhy@stanford.edu'];
+      const toEmails = ['tyler.hannasch@gmail.com', 'newton1988@gmail.com'];
       sendEmail(toEmails, defaultFromEmail, 'Feedback: ' + req.body.text, 'Email: ' + decoded.email);
 
-      connection.query('INSERT INTO feedback (text, email, school) VALUES (?, ?, ?)', [req.body.text, decoded.email, school], (err2, result) => {
+      connection.query('INSERT INTO feedback (text, email, groupId, school, type) VALUES (?, ?, ?, ?, ?)', [req.body.text, decoded.email, decoded.groupId, school, req.body.type], (err2, result) => {
         if (err2) throw err2;
         res.json({ id: result.insertId });
       });
@@ -365,8 +183,9 @@ app.post('/addProject', upload.array(), (req, res) => {
     } else {
       const title = (req.body.feedback.text) ? req.body.feedback.text : 'Blank Title';
       const school = (req.body.feedback.school) ? req.body.feedback.school : getDomain(decoded.email);
+      const groupId = (req.body.feedback.groupId) ? req.body.feedback.groupId : decoded.groupId;
 
-      connection.query('INSERT INTO projects SET ?', { title, description: 'Blank Description', votes: 0, stage: 'new', school }, (err2, result) => {
+      connection.query('INSERT INTO projects SET ?', { title, groupId, description: 'Blank Description', votes: 0, downvotes: 0, stage: 'new', school, type: req.body.feedback.type }, (err2, result) => {
         if (err2) throw err2;
         if (req.body.feedback) {
           //sendEmail(['tyler.hannasch@gmail.com'], defaultFromEmail, 'A new project has been created for your feedback', 'The next step is to get people to upvote it so it is selected for action by the department heads');
@@ -382,14 +201,25 @@ app.post('/addProject', upload.array(), (req, res) => {
 
 app.post('/addSolution', upload.array(), (req, res) => {
   jwt.verify(req.body.authorization, 'buechelejedi16', (err, decoded) => {
-
     if (err) {
       res.status(400).send('authorization failed');
     } else {
-      connection.query('INSERT INTO project_additions SET ?', { type: 'solution', votes: 0, title: req.body.title || 'Title Here', description: req.body.description || 'Description Here', project_id: req.body.project_id, school: getDomain(decoded.email), email: decoded.email }, (err, result) => {
-        if (err) throw err;
-        res.json({ id: result.insertId });
-      });
+      connection.query('INSERT INTO project_additions SET ?',
+        {
+          type: 'solution',
+          votes: 0,
+          downvotes: 0,
+          title: req.body.title || 'Title Here',
+          description: req.body.description || 'Description Here',
+          project_id: req.body.project_id,
+          school: getDomain(decoded.email),
+          groupId: decoded.groupId,
+          email: decoded.email,
+          approved: !req.body.moderatorApprovalSolutions,
+        }, (innerError, result) => {
+          if (innerError) throw innerError;
+          res.json({ id: result.insertId });
+        });
     }
   });
 });
@@ -399,7 +229,12 @@ app.post('/addSubscriber', upload.array(), (req, res) => {
     if (err) {
       res.status(400).send('authorization failed');
     } else {
-      connection.query('INSERT INTO subscriptions SET ?', { project_id: req.body.projectId, email: decoded.email, type: req.body.type }, (err2) => {
+      connection.query('INSERT INTO subscriptions SET ?',
+        {
+          project_id: req.body.projectId,
+          email: decoded.email,
+          type: req.body.type
+        }, (err2) => {
         if (err2) throw err2;
         res.sendStatus(200);
       });
@@ -413,7 +248,7 @@ app.post('/saveProjectChanges', upload.array(), (req, res) => {
     if (err) {
       res.status(400).send('authorization failed');
     } else {
-      connection.query('UPDATE projects SET votes = ?, title = ?, description = ? WHERE id= ?', [req.body.project.votes, req.body.project.title, req.body.project.description, req.body.project.id], (err) => {
+      connection.query('UPDATE projects SET votes = ?, downvotes = ?, title = ?, description = ? WHERE id= ?', [req.body.project.votes, req.body.project.downvotes, req.body.project.title, req.body.project.description, req.body.project.id], (err) => {
         if (err) throw err;
       });
       res.sendStatus(200);
@@ -426,7 +261,12 @@ var addSubscriber = function(req, res, next) {
     if (err) {
       res.status(400).send('authorization failed');
     } else {
-      connection.query('INSERT INTO subscriptions SET ?', { project_id: req.body.project_addition.id, email: decoded.email, type: 'up vote solution' }, (err2) => {
+      connection.query('INSERT INTO subscriptions SET ?',
+        {
+          project_id: req.body.project_addition.id,
+          email: decoded.email,
+          type: 'up vote solution'
+        }, (err2) => {
         if (err2) throw err2;
       });
     }
@@ -439,9 +279,17 @@ app.post('/saveProjectAdditionChanges', upload.array(), addSubscriber, (req, res
     if (err) {
       res.status(400).send('authorization failed');
     } else {
-      connection.query('UPDATE project_additions SET votes = ?, title = ?, description = ? WHERE id= ?', [req.body.project_addition.votes, req.body.project_addition.title, req.body.project_addition.description, req.body.project_addition.id], function(err) {
-        if (err) throw err;
-      });
+      connection.query('UPDATE project_additions SET votes = ?, downvotes = ?, title = ?, description = ?, approved = ? WHERE id= ?',
+        [
+          req.body.project_addition.votes,
+          req.body.project_addition.downvotes,
+          req.body.project_addition.title,
+          req.body.project_addition.description,
+          req.body.project_addition.approved,
+          req.body.project_addition.id,
+        ], (innerError) => {
+          if (innerError) throw innerError;
+        });
       res.sendStatus(200);
     }
   });
@@ -474,7 +322,6 @@ app.post('/deleteProjectAddition', upload.array(), (req, res) => {
   });
 });
 
-
 // Pull Feedback, Projects, Project Additions, Discussion Posts, and Features
 app.post('/pullFeedback', upload.array(), (req, res) => {
   jwt.verify(req.body.authorization, 'buechelejedi16', (err, decoded) => {
@@ -490,9 +337,9 @@ app.post('/pullFeedback', upload.array(), (req, res) => {
           time
             BETWEEN ? AND ?
         AND
-          school = ?`;
-      connection.query(connectionString, [req.body.startDate, req.body.endDate, getDomain(decoded.email)], (err2, rows) => {
-        if (err2) throw err2;
+          groupId = ?`;
+      connection.query(connectionString, [req.body.startDate, req.body.endDate, decoded.groupId], (err2, rows) => {
+        if (err2) res.status(400).send('Invalid Token');
         else {
           res.json(rows);
         }
@@ -508,13 +355,13 @@ app.post('/pullProjects', upload.array(), (req, res) => {
     } else {
       const connectionString = `
         SELECT
-          id, title, votes, description, department, stage
+          id, title, votes, downvotes, description, department, stage, type
         FROM
           projects
         WHERE
-          school=?`;
-      connection.query(connectionString, [getDomain(decoded.email)], (err2, rows) => {
-        if (err2) throw err2;
+          groupId=?`;
+      connection.query(connectionString, [decoded.groupId], (err2, rows) => {
+        if (err2 || !decoded.groupId) res.status(400).send('Invalid Token');
         else res.send(rows);
       });
     }
@@ -528,34 +375,14 @@ app.post('/pullProjectAdditions', upload.array(), (req, res) => {
     } else {
       const connectionString = `
         SELECT
-          id, type, votes, title, description, project_id
+          id, type, votes, downvotes, title, description, project_id, approved
         FROM
           project_additions
         WHERE
-          school=?`;
-      connection.query(connectionString, [getDomain(decoded.email)], (err2, rows) => {
-        if (err2) throw err2;
+          groupId=?`;
+      connection.query(connectionString, [decoded.groupId], (err2, rows) => {
+        if (err2) res.status(400).send('Invalid Token');
         else res.send(rows);
-      });
-    }
-  });
-});
-
-app.post('/pullDiscussionPosts', upload.array(), (req, res) => {
-  jwt.verify(req.body.authorization, 'buechelejedi16', (err) => {
-    if (err) {
-      res.status(400).send('authorization failed');
-    } else {
-      const connectionString = `
-        SELECT
-          id, point, counter_point, project_addition_id
-        FROM
-          discussion_posts`;
-      connection.query(connectionString, (err2, rows) => {
-        if (err) throw err;
-        else {
-          res.send(rows);
-        }
       });
     }
   });
@@ -568,13 +395,19 @@ app.post('/pullFeatures', upload.array(), (req, res) => {
     } else {
       const connectionString = `
         SELECT
-          moderator_approval as moderatorApproval, show_status as showStatus
+          moderator_approval AS moderatorApproval,
+          moderator_approval_solutions AS moderatorApprovalSolutions,
+          show_status AS showStatus,
+          enable_new_feedback AS enableNewFeedback,
+          positive_feedback_box AS positiveFeedbackBox,
+          ? AS domain,
+          ? AS email
         FROM
           features
         WHERE
-          school=?`;
-      connection.query(connectionString, [getDomain(decoded.email)], (err2, rows) => {
-        if (err2) throw err2;
+          id=?`;
+      connection.query(connectionString, [getDomain(decoded.email), decoded.email, decoded.groupId], (err2, rows) => {
+        if (err2) res.status(400).send('Invalid Token');
         else res.send(rows);
       });
     }
